backend/requirements.txt
Flask>=2.2.5
flask-cors>=3.0.10
pymongo[srv]>=4.3.3
flask-jwt-extended>=4.4.4
python-dotenv>=1.0.0
werkzeug>=2.2.0
cloudinary>=1.29.0
requests>=2.31.0
gunicorn>=20.1.0
apscheduler>=3.10.1
certifi
